Add timing to all balboa-http endpoints that interact with cassandra
