+++ balboa-http/src/main/scala/com/socrata/balboa/server/MainServlet.scala
@@ -1,19 +1,24 @@
 package com.socrata.balboa.server
 
+import java.util.regex.Pattern
 import javax.servlet.http.HttpServletRequest
 
 import com.socrata.balboa.BuildInfo
-import com.socrata.balboa.metrics.data.BalboaFastFailCheck
-import com.socrata.balboa.server.rest.EntitiesRest
+import com.socrata.balboa.metrics.data.{BalboaFastFailCheck, DataStoreFactory}
 import com.socrata.balboa.server.ResponseWithType.json
 import com.typesafe.scalalogging.slf4j.StrictLogging
+import org.codehaus.jackson.map.annotate.JsonSerialize
+import org.codehaus.jackson.map.{ObjectMapper, SerializationConfig}
+import org.eclipse.jetty.http.HttpStatus.BAD_REQUEST_400
 import org.json4s.{DefaultFormats, Formats}
-import org.scalatra.{InternalServerError, Ok, ScalatraServlet}
+import org.scalatra._
 import org.scalatra.json.JacksonJsonSupport
+import org.scalatra.metrics.MetricsSupport
 
 import scala.collection.JavaConverters._
 
 class MainServlet extends ScalatraServlet
+    with MetricsSupport
     with ClientCounter
     with StrictLogging
     with JacksonJsonSupport
@@ -22,6 +27,7 @@ class MainServlet extends ScalatraServlet
 
   override protected implicit val jsonFormats: Formats = DefaultFormats
 
+  val dataStore = DataStoreFactory.get
   val versionString = BuildInfo.toJson
 
   def getAccepts(req: HttpServletRequest): Seq[String] = {
@@ -37,9 +43,32 @@ class MainServlet extends ScalatraServlet
     }
   }
 
-  get("/entities*") {
-    val response = EntitiesRest(params)
-    contentType = response.contentType
-    response.result
+  get("/entities*")(getEntities)
+  def getEntities: ActionResult = {
+    contentType = json
+
+    val predicate: (String => Boolean) =
+      params.get("filter").map(Pattern.compile)
+        .map { pattern => (str: String) => pattern.matcher(str).matches }.getOrElse(_ => true)
+
+    timer("entities queries")({
+      val it = dataStore.entities().asScala.filter(predicate)
+      // backwards-compatibility: -1 == no limit
+      val limitString = params.getOrElse("limit", "-1")
+      val limit = try {
+        limitString.toInt
+      } catch {
+        case err: NumberFormatException =>
+          return BadRequest(Error(BAD_REQUEST_400, "Unable to parse limit as int"))
+      }
+
+      val limitedIt = if (limit != -1) it.take(limit) else it
+
+      val mapper = new ObjectMapper
+      mapper.configure(SerializationConfig.Feature.INDENT_OUTPUT, true)
+      mapper.getSerializationConfig.withSerializationInclusion(JsonSerialize.Inclusion.NON_NULL)
+
+      Ok(mapper.writeValueAsString(limitedIt.asJava))
+    }).call()
   }
 }
