+++ balboa-http/src/main/scala/com/socrata/balboa/server/MetricsServlet.scala
@@ -13,7 +13,7 @@ import com.typesafe.scalalogging.slf4j.StrictLogging
 import org.codehaus.jackson.map.annotate.JsonSerialize
 import org.codehaus.jackson.map.{ObjectMapper, SerializationConfig}
 import org.scalatra.metrics.MetricsSupport
-import org.scalatra.{NoContent, Ok}
+import org.scalatra.{ActionResult, NoContent, Ok}
 
 class MetricsServlet extends JacksonJsonServlet
     with MetricsSupport
@@ -35,7 +35,7 @@ class MetricsServlet extends JacksonJsonServlet
 
   // Match paths like /metrics/:entityId and /metrics/:entityId/whatever
   get("""^\/([^\/]+).*""".r)(getMetrics)
-  def getMetrics: Any = {
+  def getMetrics: ActionResult = {
     val entityId = params("captures")
 
     val period = params.get(PeriodKey).map(Extractable[Period].extract) match {
@@ -80,7 +80,7 @@ class MetricsServlet extends JacksonJsonServlet
   }
 
   get("/:entityId/range*")(getRange)
-  def getRange: Any = {
+  def getRange: ActionResult = {
     val entityId = params("entityId")
 
     val start = params.getOrElse(StartKey, {
@@ -105,7 +105,7 @@ class MetricsServlet extends JacksonJsonServlet
 
     val mediaType = bestMediaType(getAccepts(request), json, protobuf).getOrElse({
       contentType = json
-      return unacceptable
+      return unacceptable.result
     })
 
     timer("range queries")({
@@ -122,7 +122,7 @@ class MetricsServlet extends JacksonJsonServlet
   }
 
   get("/:entityId/series*")(getSeries)
-  def getSeries: Any = {
+  def getSeries: ActionResult = {
     val entityId = params("entityId")
 
     val period = params.get(PeriodKey).map(Extractable[Period].extract) match {
@@ -149,12 +149,12 @@ class MetricsServlet extends JacksonJsonServlet
     })
     val endDate = ServiceUtils.parseDate(end).getOrElse({
       contentType = json
-      return malformedDate(end)
+      return malformedDate(end).result
     })
 
     bestMediaType(getAccepts(request), json).getOrElse({
       contentType = json
-      return unacceptable
+      return unacceptable.result
     })
 
     timer("series queries")({
@@ -166,7 +166,7 @@ class MetricsServlet extends JacksonJsonServlet
   }
 
   post("/:entityId")(postMetrics())
-  def postMetrics(): Any = {
+  def postMetrics(): ActionResult = {
     // Note: The `contentType` is used to determine how the response is to be
     // interpreted. For some unclear reason, it must be set -before- calling
     // extractOpt for the body of the request in order for it to take effect on
@@ -194,7 +194,10 @@ class MetricsServlet extends JacksonJsonServlet
       metrics.put(name, new Metric(recordType, metric.value))
     }
 
-    dataStore.persist(entityId, entity.timestamp, metrics)
+    timer("metric entity post")({
+      dataStore.persist(entityId, entity.timestamp, metrics)
+    }).call()
+
     NoContent()
   }
